Move color information out of SpriteLibrary and into generic UI
