+++ src/UI/Art/SpriteArtist/SpriteLibrary.java
@@ -14,6 +14,7 @@ import CommandingOfficers.Commander;
 import CommandingOfficers.CommanderLibrary;
 import Terrain.Location;
 import Terrain.TerrainType;
+import UI.UIUtils;
 import Units.Unit;
 import Units.UnitModel;
 
@@ -24,44 +25,6 @@ public class SpriteLibrary
 {
   // This is the physical size of a single map square in pixels.
   public static final int baseSpriteSize = 16;
-  
-  public static final String DEFAULT_FACTION_NAME = "Thorn";
-
-  // Define extra colors as needed.
-  private static final Color PURPLE = new Color(231, 123, 255 );
-
-  // Map Building colors.
-  public static final Color[] defaultMapColors = { new Color(40, 40, 40), new Color(70, 70, 70), new Color(110, 110, 110), new Color(160, 160, 160),
-      new Color(200, 200, 200), new Color(231, 231, 231) };
-  private static Color[] pinkMapBuildingColors = { new Color(142, 26, 26), new Color(255, 219, 74), new Color(190, 90, 90), new Color(240, 140, 140),
-      new Color(250, 190, 190), new Color(255, 245, 245) };
-  private static Color[] cyanMapBuildingColors = { new Color(0, 105, 105), new Color(255, 219, 74), new Color(77, 157, 157), new Color(130, 200, 200),
-      new Color(200, 230, 230), new Color(245, 255, 255) };
-  private static Color[] orangeMapBuildingColors = { new Color(130, 56, 0), new Color(255, 237, 29), new Color(139, 77, 20), new Color(231, 139, 41),
-      new Color(243, 186, 121), new Color(255, 234, 204) };
-  private static Color[] purpleMapBuildingColors = { new Color(90, 14, 99), new Color(255, 207, 95), new Color(133, 65, 130), new Color(174, 115, 189),
-    new Color(222, 171, 240), new Color(255, 231, 255) };
-
-  // Map Unit colors.
-  private static Color[] pinkMapUnitColors = { new Color(142, 26, 26), new Color(199, 62, 62), new Color(248, 100, 100), new Color(255, 136, 136),
-      new Color(255, 175, 175), new Color(255, 201, 201) };
-  private static Color[] cyanMapUnitColors = { new Color(0, 105, 105), new Color(0, 170, 170), new Color(0, 215, 215), new Color(0, 245, 245),
-      new Color(121, 255, 255), new Color(195, 255, 255), };
-  private static Color[] orangeMapUnitColors = { new Color(130, 56, 0), new Color(204, 103, 7), new Color(245, 130, 14), new Color(255, 160, 30),
-      new Color(255, 186, 60), new Color(255, 225, 142), };
-  private static Color[] purpleMapUnitColors = { new Color(90, 14, 99), new Color(132, 41, 148), new Color(181, 62, 198), new Color(201, 98, 223),
-    new Color(231, 123, 255), new Color(243, 180, 255), };
-
-  private static HashMap<Color, ColorPalette> buildingColorPalettes;
-  private static HashMap<Color, ColorPalette> mapUnitColorPalettes;
-  private static HashMap<Color, String> colorNames;
-  private static ArrayList<String> factionNames;
-
-  public static Color[] getCOColors()
-  {
-    initCosmetics();
-    return mapUnitColorPalettes.keySet().toArray(new Color[0]);
-  }
 
   // TODO: Account for weather?
   private static HashMap<SpriteSetKey, TerrainSpriteSet> spriteSetMap = new HashMap<SpriteSetKey, TerrainSpriteSet>();
@@ -239,7 +202,7 @@ public class SpriteLibrary
     // If this tile is owned by someone, fly their colors.
     if( spriteKey.commanderKey != null )
     {
-      ss.colorize(defaultMapColors, getBuildingColors(spriteKey.commanderKey.myColor).paletteColors);
+      ss.colorize(UIUtils.defaultMapColors, UIUtils.getBuildingColors(spriteKey.commanderKey.myColor).paletteColors);
     }
     spriteSetMap.put(spriteKey, ss);
   }
@@ -277,107 +240,6 @@ public class SpriteLibrary
     }
 
     return moveCursorArrowSprite;
-  }
-  
-  /**
-   * Parses the available palettes and faction unit images.
-   * Both of the above are parsed from res/unit/faction.
-   * For palettes:
-   *   Reads in the first 6 colors on each row.
-   *   Bottom colors are buildings, top are units.
-   *   The key color is the last color on the top row.
-   * For factions:
-   *   The name of the folder is simply collected; no further work is done until units are drawn.
-   */
-  private static void initCosmetics()
-  {
-    if (null == mapUnitColorPalettes )
-    {
-      buildingColorPalettes = new HashMap<Color, ColorPalette>();
-      mapUnitColorPalettes = new HashMap<Color, ColorPalette>();
-      colorNames = new HashMap<Color, String>();
-      factionNames = new ArrayList<String>();
-
-      // Create a mapping of game colors to the fine-tuned colors that will be used for map sprites.
-      buildingColorPalettes.put(Color.PINK, new ColorPalette(pinkMapBuildingColors));
-      buildingColorPalettes.put(Color.CYAN, new ColorPalette(cyanMapBuildingColors));
-      buildingColorPalettes.put(Color.ORANGE, new ColorPalette(orangeMapBuildingColors));
-      buildingColorPalettes.put(PURPLE, new ColorPalette(purpleMapBuildingColors));
-
-      mapUnitColorPalettes.put(Color.PINK, new ColorPalette(pinkMapUnitColors));
-      mapUnitColorPalettes.put(Color.CYAN, new ColorPalette(cyanMapUnitColors));
-      mapUnitColorPalettes.put(Color.ORANGE, new ColorPalette(orangeMapUnitColors));
-      mapUnitColorPalettes.put(PURPLE, new ColorPalette(purpleMapUnitColors));
-
-      // Throw some color names in there for the defaults
-      // toString() is not user-friendly
-      colorNames.put(Color.PINK, "salmon");
-      colorNames.put(Color.CYAN, "cyan");
-      colorNames.put(Color.ORANGE, "citrus");
-      colorNames.put(PURPLE, "sparking");
-
-      // We want to be able to use the normal units, as well as any others
-      factionNames.add(DEFAULT_FACTION_NAME);
-
-      final File folder = new File("res/unit/faction");
-
-      if (folder.canRead())
-      {
-        for( final File fileEntry : folder.listFiles() )
-        {
-          // If it's a file, we assume it's a palette
-          if( !fileEntry.isDirectory() )
-          {
-            String colorName = fileEntry.getAbsolutePath();
-            if( colorName.contains(".png") )
-            {
-              BufferedImage bi = loadSpriteSheetFile(colorName);
-              // Grab the last color on the first row as our "banner" color
-              Color key = new Color(bi.getRGB(bi.getWidth() - 1, 0));
-              Color[] cUnits = new Color[defaultMapColors.length];
-              Color[] cStructs = new Color[defaultMapColors.length];
-              for( int i = 0; i < defaultMapColors.length; i++ )
-              {
-                cUnits[i] = new Color(bi.getRGB(i, 0));
-                cStructs[i] = new Color(bi.getRGB(i, 1));
-              }
-              buildingColorPalettes.put(key, new ColorPalette(cStructs));
-              mapUnitColorPalettes.put(key, new ColorPalette(cUnits));
-              colorNames.put(key, fileEntry.getName().replace(".png", ""));
-            }
-          }
-          // If it's a directory, we assume it's a set of map sprites, i.e. a faction.
-          else if(!fileEntry.getName().endsWith(DEFAULT_FACTION_NAME)) // However, we don't wanna add our default twice
-          {
-            factionNames.add(fileEntry.getName());
-          }
-        }
-      }
-    }
-  }
-
-  public static ColorPalette getBuildingColors(Color colorKey)
-  {
-    initCosmetics();
-    return buildingColorPalettes.get(colorKey);
-  }
-
-  public static ColorPalette getMapUnitColors(Color colorKey)
-  {
-    initCosmetics();
-    return mapUnitColorPalettes.get(colorKey);
-  }
-
-  public static String getColorName(Color colorKey)
-  {
-    initCosmetics();
-    return colorNames.get(colorKey);
-  }
-
-  public static String[] getFactionNames()
-  {
-    initCosmetics();
-    return factionNames.toArray(new String[0]);
   }
 
   private static class SpriteSetKey
@@ -466,7 +328,7 @@ public class SpriteLibrary
     System.out.println("creating " + key.unitTypeKey.toString() + " spriteset for CO " + key.commanderKey.myColor.toString() + " in faction " + faction);
     String filestr = getMapUnitSpriteFilename(key.unitTypeKey, faction);
     UnitSpriteSet spriteSet = new UnitSpriteSet(loadSpriteSheetFile(filestr), baseSpriteSize, baseSpriteSize,
-        getMapUnitColors(key.commanderKey.myColor));
+        UIUtils.getMapUnitColors(key.commanderKey.myColor));
     mapUnitSpriteSetMap.put(key, spriteSet);
   }
 
@@ -709,7 +571,7 @@ public class SpriteLibrary
 
       // If we don't already have this overlay, go load and store it.
       Sprite overlay = new Sprite(loadSpriteSheetFile("res/ui/co_overlay.png"), OVERLAY_WIDTH, OVERLAY_HEIGHT);
-      overlay.colorize(defaultMapColors, mapUnitColorPalettes.get(co.myColor).paletteColors);
+      overlay.colorize(UIUtils.defaultMapColors, UIUtils.getMapUnitColors(co.myColor).paletteColors);
 
       // Draw the Commander's mug on top of the overlay.
       BufferedImage coMug = getCommanderSprites(co.coInfo.name).eyes;
@@ -733,7 +595,7 @@ public class SpriteLibrary
     Graphics barGfx = bar.getGraphics();
 
     // Get the CO's colors
-    Color[] palette = SpriteLibrary.mapUnitColorPalettes.get(co.myColor).paletteColors;
+    Color[] palette = UIUtils.getMapUnitColors(co.myColor).paletteColors;
 
     // Draw the bar
     barGfx.setColor(Color.BLACK);               // Outside edge
@@ -760,7 +622,7 @@ public class SpriteLibrary
       // Image should be: empty bar, full bar, empty point, 1/3 point, 2/3 point, full point, large point.
       // Image should be: empty point, 1/3 point, 2/3 point, full point, large point.
       Sprite overlay = new Sprite(loadSpriteSheetFile("res/ui/powerbar_pieces.png"), POWERBAR_FRAME_WIDTH, POWERBAR_FRAME_HEIGHT);
-      overlay.colorize(defaultMapColors, mapUnitColorPalettes.get(co.myColor).paletteColors);
+      overlay.colorize(UIUtils.defaultMapColors, UIUtils.getMapUnitColors(co.myColor).paletteColors);
 
       coPowerBarPieces.put(co, overlay);
     }
