+++ src/UI/COSetupController.java
@@ -42,7 +42,7 @@ public class COSetupController implements IController
     for(int co = 0; co < numCos; ++co)
     {
       // Set up our option selection framework
-      coSelectors[co] = new COSetupInfo(numCos, co, CommanderLibrary.getCommanderList(), SpriteLibrary.getCOColors(), SpriteLibrary.getFactionNames(), AILibrary.getAIList());
+      coSelectors[co] = new COSetupInfo(numCos, co, CommanderLibrary.getCommanderList(), UIUtils.getCOColors(), UIUtils.getFactionNames(), AILibrary.getAIList());
     }
   }
